requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
fake-useragent==0.1.11
Flask==2.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
psycopg2==2.9.1
PyYAML==5.4.1
requests==2.26.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.6.0
